% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISamplR.R
\docType{package}
\name{AISamplR}
\alias{AISamplR}
\alias{AISamplR-package}
\title{AIS: A package for Adaptive Importance Sampling.}
\description{
The AISamplR provides three function for Adaptive Importance Sampling:
  \code{\link{lais}}, \code{\link{apis}} and \code{link{pmc}}.
  
For a complete description of the three methods,
please see the following references:
}
