% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_logdenom_bytable}
\alias{compute_logdenom_bytable}
\title{Computes the logarithm of the denominator of importance sampling weights.}
\usage{
compute_logdenom_bytable(x, mu, sigma2, D, T, N, M)
}
\arguments{
\item{x}{A numerical array of dimension D x T x N x M. 
It is the result of the sampling step in the adapative importance sampling scheme.}

\item{mu}{A numerical array of dimension D x T x N.
It is the result of the adaptaion step in the adapative importance sampling scheme.}

\item{sigma2}{A numerical matrix of length D.
It provides the variance of each dimension of the gaussian proposal distribution
used in the sampling step.}

\item{D}{A integer providing the dimension of the space of inputs.}

\item{T}{A integer providing the number iterations used in the adaptation step.}

\item{N}{A integer providing the number of proposal distribution chains used.}

\item{M}{A integer providing the number of samples drawn from each proposal.}
}
\value{
A array of dimension T x N x M with the logarithm of the 
denominator needed to compute the importance sampling weights.
}
\description{
\code{\link{compute_logdenom_bybox}}, code{\link{compute_logdenom_byrow}} and \code{\link{compute_logdenom_bytable}}
offers three different way to compute the logdenominator of the importance sampling weights.
}
\details{
The weighting step can be performed in different way depending on 
how we considered the samples drawn at the sampling step.
We can think that one sample was drawn from its own corresponding proposal distribution 
or that one sample was drawn from a mixture of  proposal distributions. 
In this package, three methods to compute the denominator of the weights are available:
\itemize{
  \item \code{\link{compute_logdenom_bybox}} where we considered that on sample x_\{t,n,m\} 
  is drawn from an unique 
  proposal distribution with location parameter mu_\{t,n\}.
  \item \code{\link{compute_logdenom_byrow}} where we considered that on sample  x_\{t,n,m\} 
  is drawn from an equiprobable mixture
  of all proposal distributions at time t.
  \item \code{\link{compute_logdenom_bytable}} where we considered that on sample x_\{t,n,m\} 
  is drawn from an equiprobable mixture
  of all available proposal distributions, for all times t = 1,..., T and 
  chains of proposal n = 1,..., N.
}
}
\examples{
D <- 1
T <- 10
N <- 2
M <- 3
pmc_lp_r <- pmc(dexp,
    mu = matrix(rnorm(D*N, mean = 1, sd = 1), nrow = D, ncol = N),
    sig2_adapt = rep(1, D), sig2_samp = rep(1, D),
    compute_logdenom = compute_logdenom_byrow,
    N = N, T = T, M = M)
with(pmc_lp6_r, compute_expectation(x, weight)) # theorical value: ~ [-1.09, 0]
}
