% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgl_plot}
\alias{rgl_plot}
\title{3D scatterplot with rgl.}
\usage{
rgl_plot(x, y, z, ...)
}
\arguments{
\item{x}{a vector with the x-axis coordinates}

\item{y}{a vector with the y-axis coordinates}

\item{z}{a vector with the z-axis coordinates}

\item{...}{additional arguments for the function \code{\link[rgl]{plot3d}}.}
}
\description{
\code{rgl_plot} makes a 3D scatterplot with a rainbow color scale.
}
\details{
The function is a simple wrapper around \code{\link[rgl]{plot3d}}.
It produces a 3D scatterplot with a rainbow color scale for the z-axis.
}
\examples{

# draw samples from the "banana shaped distribution" defined by the loglikelihood lposterior_6.

lposterior_6 <- function(x){
  x1 <- x[1]
  x2 <- x[2]
  logtarget <-  -1/32 * (4 - 10 * x1 - x2^2)^2 - x1^2/50 - x2^2/50
}
D <- 2
T <- 100
N <- 50
M <- 4

lais_lp6_r <- lais(lposterior_6,
   mu = matrix(rnorm(D * N, sd = 3), nrow = D, ncol = N),
   sig2_adapt = rep(1, D), sig2_samp = rep(1, D),
   compute_logdenom = compute_logdenom_byrow,
   N = N, T = T, M = M)
   
with(lais_lp6_r, rgl_plot(x[1,,,], x[2,,,], exp(loglik)))
with(lais_lp6_r, rgl_plot(x[1,,,], x[2,,,], sqrt(weight)))

}
